+++ lib/linker.dart
@@ -30,6 +30,9 @@ void link(GenerationOptions options, Iterable<FileGenerator> files) {
     for (var s in f.serviceGenerators) {
       s.resolve(ctx);
     }
+    for (var s in f.grpcGenerators) {
+      s.resolve(ctx);
+    }
   }
 }