+++ lib/protoc.dart
@@ -3,6 +3,7 @@ library protoc;
 import 'dart:async';
 import 'dart:io';
 
+import 'package:dart_style/dart_style.dart';
 import 'package:protobuf/mixins_meta.dart';
 import 'package:protobuf/protobuf.dart';
 import 'package:path/path.dart' as path;
@@ -21,6 +22,7 @@ part 'code_generator.dart';
 part 'enum_generator.dart';
 part 'extension_generator.dart';
 part 'file_generator.dart';
+part 'grpc_generator.dart';
 part 'linker.dart';
 part 'message_generator.dart';
 part 'options.dart';